backend/requirements.txt
anthropic==0.39.0
fastapi==0.117.1
openai==1.109.1
pydantic==2.11.9
python-dotenv==1.1.1
requests==2.32.5
uvicorn==0.37.0